requirements.txt
# swarm-ai==0.1.0 directly download from the openai swarm-ai repo
openai==1.12.0
pdfminer.six==20221105
python-dotenv==1.0.0
rich==13.7.0
streamlit==1.32.0
streamlit-extras==0.4.0
streamlit-option-menu==0.3.12
google-generativeai
